// interfaces/ads.ts
// Generated by https://quicktype.io

export interface Ads {
    _id:       string;
    mainImage: MainImage;
    name:      string;
    socials:   Socials;
}

export interface MainImage {
    _type: string;
    asset: Asset;
}

export interface Asset {
    _ref:  string;
    _type: string;
}

export interface Socials {
    instagram?: string;
    facebook?: string;
    whatsapp?: string;
}